Add UI class and Storage class This is to achieve better OOP design.
